@@ -26,40 +26,39 @@
   },
   "preferGlobal": true,
   "scripts": {
-    "test": "tslint -c ./tslint.json -p ./tsconfig.json",
+    "lint": "tslint -c ./tslint.json -p ./tsconfig.json",
     "build": "tsc -p ./tsconfig.json",
-    "release": "standard-version",
-    "prepublish": "npm run test && npm run build"
+    "release": "standard-version"
   },
   "bin": {
     "ssc": "bin/ssc"
   },
   "dependencies": {
-    "chalk": "1.1.3",
-    "commander": "2.9.0",
-    "deepmerge": "1.3.2",
-    "fs-extra": "3.0.1",
+    "chalk": "2.0.1",
+    "commander": "2.11.0",
+    "deepmerge": "1.5.0",
+    "fs-extra": "4.0.0",
     "glob": "7.1.2",
-    "inquirer": "3.1.0",
+    "inquirer": "3.2.1",
     "mssql": "4.0.4",
     "multimatch": "2.1.0",
     "tedious": "2.0.0",
-    "ts-util-is": "1.0.3",
+    "ts-util-is": "1.1.2",
     "xml2js": "0.4.17"
   },
   "devDependencies": {
     "@types/chalk": "0.4.31",
     "@types/commander": "2.9.1",
     "@types/deepmerge": "1.3.1",
-    "@types/fs-extra": "3.0.3",
+    "@types/fs-extra": "4.0.0",
     "@types/glob": "5.0.30",
-    "@types/inquirer": "0.0.34",
-    "@types/mssql": "4.0.1",
+    "@types/inquirer": "0.0.35",
+    "@types/mssql": "4.0.5",
     "@types/multimatch": "2.1.1",
-    "@types/xml2js": "0.0.33",
-    "standard-version": "4.0.0",
-    "tslint": "5.4.3",
-    "tslint-config-unicorn-science": "1.0.1",
-    "typescript": "2.3.4"
+    "@types/xml2js": "0.4.0",
+    "standard-version": "4.2.0",
+    "tslint": "5.5.0",
+    "tslint-config-unicorn-science": "1.1.7",
+    "typescript": "2.4.2"
   }
 }
